+++ PLPlugins/PLParticleGroups/src/PGRain.cpp
@@ -124,25 +124,26 @@ void PGRain::InitParticle(Particle &cParticle) const
 		float fVelLength = MediumSpeed/cParticle.vVelocity.GetLength();
 		cParticle.vVelocity.x *= fVelLength;
 		cParticle.vVelocity.z *= fVelLength;
-		fVelLength += (float)(0.2*fVelLength);
+		fVelLength += static_cast<float>(0.2*fVelLength);
 		cParticle.vVelocity.y *= fVelLength;
 	} else {
 		float fVelLength = 1/cParticle.vVelocity.GetLength();
 		cParticle.vVelocity.x *= fVelLength*MediumSpeed/5;
 		cParticle.vVelocity.y *= fVelLength*MediumSpeed/5;
 		cParticle.vVelocity.z *= (fVelLength*MediumSpeed);
-		if (cParticle.vVelocity.z > 1) cParticle.vVelocity.z = -cParticle.vVelocity.z;
+		if (cParticle.vVelocity.z > 1)
+			cParticle.vVelocity.z = -cParticle.vVelocity.z;
 	}
 	cParticle.fEnergy = Math::GetRandNegFloat()*EnergyRange + EnergyMin;
 
 	float fColor = Math::GetRandFloat()*0.5f;
-	cParticle.vColor[0] = ((float)Brightness/255)-fColor;
-	cParticle.vColor[1] = ((float)Brightness/255)-fColor;
-	cParticle.vColor[2] = ((float)Brightness/255)-fColor;
-	cParticle.vColor[3] = ((float)cParticle.fEnergy/255);
+	cParticle.vColor[0] = (static_cast<float>(Brightness)/255.0f)-fColor;
+	cParticle.vColor[1] = (static_cast<float>(Brightness)/255.0f)-fColor;
+	cParticle.vColor[2] = (static_cast<float>(Brightness)/255.0f)-fColor;
+	cParticle.vColor[3] = cParticle.fEnergy/255.0f;
 
-	cParticle.fSize    = (float)(MediumSize + Math::GetRandNegFloat()*MediumSize*0.1)/10;
-	cParticle.fCustom1 = (float)(cParticle.fSize*1.9);
+	cParticle.fSize    = (static_cast<float>(MediumSize) + Math::GetRandNegFloat()*static_cast<float>(MediumSize)*0.1f)/10.0f;
+	cParticle.fCustom1 = cParticle.fSize*1.9f;
 
 	cParticle.bDistorted	= true;
 	cParticle.vDistortion.x = 0.0f;
@@ -164,8 +165,10 @@ void PGRain::NotifyUpdate()
 		if (!(System::GetInstance()->GetMicroseconds() % BuildPerSec)) {
 			while (Math::GetRand() % 5) {
 				Particle *pParticle = AddParticle();
-				if (pParticle) InitParticle(*pParticle);
-				else		   break;
+				if (pParticle)
+					InitParticle(*pParticle);
+				else
+					break;
 			}
 		}
 
@@ -193,7 +196,8 @@ void PGRain::NotifyUpdate()
 				cParticle.fSize   -= fTimeDiff/2;
 				if (cParticle.fEnergy <= 0 || cParticle.fSize <= 0)
 					InitParticle(cParticle);
-				else cParticle.vColor[3] = cParticle.fEnergy/255;
+				else
+					cParticle.vColor[3] = cParticle.fEnergy/255;
 			}
 		}
